Better telemetry flush handling; Bugfix flush script spawning
